requirements.txt
aiogram==2.25.1
python-dotenv==0.21.1
SQLAlchemy==2.0.4
colorlog
matplotlib
requests